# Switch matmul default CPU codegen to use TileFuseAndVectorize pipeline.