@MaheshRavishankar 2021-12-08
@bjacob I think the `vector.contract` to inline ASM work you have been doing is on the old path (which I think is now switched over to use the new path). You should be able to adapt to use this path. Please get in touch with @hanhanW if you have issues.